+++ stack_configuration.go
@@ -14,8 +14,6 @@ import (
 
 // StackConfigurations describes all the stacks configurations-related methods that the
 // HCP Terraform API supports.
-// NOTE WELL: This is a beta feature and is subject to change until noted otherwise in the
-// release notes.
 type StackConfigurations interface {
 	// CreateAndUpload packages and uploads the specified Terraform Stacks
 	// configuration files in association with a Stack.
@@ -28,7 +26,7 @@ type StackConfigurations interface {
 	Read(ctx context.Context, id string) (*StackConfiguration, error)
 
 	// ListStackConfigurations returns a list of stack configurations for a stack.
-	List(ctx context.Context, stackID string, options *StackConfigurationListOptions) (*StackConfigurationList, error)
+	List(ctx context.Context, stackID string, opts *StackConfigurationListOptions) (*StackConfigurationList, error)
 
 	// JSONSchemas returns a byte slice of the JSON schema for the stack configuration.
 	JSONSchemas(ctx context.Context, stackConfigurationID string) ([]byte, error)
@@ -42,6 +40,9 @@ type StackConfigurations interface {
 	// stack configuration as it progresses, until that status is "<status>",
 	// "errored", "canceled".
 	AwaitStatus(ctx context.Context, stackConfigurationID string, status StackConfigurationStatus) <-chan WaitForStatusResult
+
+	// Diagnostics returns the diagnostics for this stack configuration.
+	Diagnostics(ctx context.Context, stackConfigurationID string) (*StackDiagnosticsList, error)
 }
 
 type StackConfigurationStatus string
@@ -253,4 +254,20 @@ func (s stackConfigurations) pollForUploadURL(ctx context.Context, stackConfigur
 // responsibility to ensure the raw content is a valid Terraform configuration.
 func (s stackConfigurations) UploadTarGzip(ctx context.Context, uploadURL string, archive io.Reader) error {
 	return s.client.doForeignPUTRequest(ctx, uploadURL, archive)
+}
+
+// Diagnostics returns the diagnostics for this stack configuration.
+func (s stackConfigurations) Diagnostics(ctx context.Context, stackConfigurationID string) (*StackDiagnosticsList, error) {
+	req, err := s.client.NewRequest("GET", fmt.Sprintf("stack-configurations/%s/stack-diagnostics", url.PathEscape(stackConfigurationID)), nil)
+
+	if err != nil {
+		return nil, err
+	}
+
+	diagnostics := &StackDiagnosticsList{}
+	err = req.Do(ctx, diagnostics)
+	if err != nil {
+		return nil, err
+	}
+	return diagnostics, nil
 }