+++ stack_deployment_steps_integration_test.go
@@ -5,6 +5,7 @@ package tfe
 
 import (
 	"context"
+	"io"
 	"testing"
 	"time"
 
@@ -237,7 +238,7 @@ func pollStackDeploymentStepStatus(t *testing.T, ctx context.Context, client *Cl
 	defer cancel()
 
 	deadline, _ := ctx.Deadline()
-	t.Logf("Polling stack deployment step %q for change in status with deadline of %s", stackDeploymentStepID, deadline)
+	t.Logf("Polling stack deployment step %q for change in status to %s with deadline of %s", stackDeploymentStepID, status, deadline)
 
 	ticker := time.NewTicker(2 * time.Second)
 	defer ticker.Stop()
@@ -262,4 +263,87 @@ func pollStackDeploymentStepStatus(t *testing.T, ctx context.Context, client *Cl
 	}
 
 	return
+}
+
+func TestStackDeploymentStepsDiagnosticsArtifacts(t *testing.T) {
+	skipUnlessBeta(t)
+
+	client := testClient(t)
+	ctx := context.Background()
+
+	orgTest, orgTestCleanup := createOrganization(t, client)
+	t.Cleanup(orgTestCleanup)
+
+	oauthClient, cleanup := createOAuthClient(t, client, orgTest, nil)
+	t.Cleanup(cleanup)
+
+	stack, err := client.Stacks.Create(ctx, StackCreateOptions{
+
+		Project: orgTest.DefaultProject,
+		Name:    "test-stack",
+
+		VCSRepo: &StackVCSRepoOptions{
+			Identifier:   "hashicorp-guides/pet-nulls-stack",
+			OAuthTokenID: oauthClient.OAuthTokens[0].ID,
+			Branch:       "main",
+		},
+	})
+	require.NoError(t, err)
+	require.NotNil(t, stack)
+
+	stackUpdated, err := client.Stacks.FetchLatestFromVcs(ctx, stack.ID)
+	require.NoError(t, err)
+	require.NotNil(t, stackUpdated)
+
+	stackUpdated = pollStackDeploymentGroups(t, ctx, client, stackUpdated.ID)
+	require.NotNil(t, stackUpdated.LatestStackConfiguration)
+
+	stackDeploymentGroups, err := client.StackDeploymentGroups.List(ctx, stackUpdated.LatestStackConfiguration.ID, nil)
+	require.NoError(t, err)
+	require.NotEmpty(t, stackDeploymentGroups)
+
+	sdg := stackDeploymentGroups.Items[0]
+
+	stackDeploymentRuns, err := client.StackDeploymentRuns.List(ctx, sdg.ID, nil)
+	require.NoError(t, err)
+	require.NotEmpty(t, stackDeploymentRuns)
+
+	sdr := stackDeploymentRuns.Items[0]
+	steps, err := client.StackDeploymentSteps.List(ctx, sdr.ID, nil)
+	assert.NoError(t, err)
+	assert.NotEmpty(t, steps)
+
+	step := steps.Items[0]
+	step = pollStackDeploymentStepStatus(t, ctx, client, step.ID, "pending_operator")
+	require.NotNil(t, step)
+
+	t.Run("Diagnostics with valid ID", func(t *testing.T) {
+		sds, err := client.StackDeploymentSteps.Diagnostics(ctx, step.ID)
+		assert.NoError(t, err)
+		assert.NotEmpty(t, sds)
+	})
+
+	t.Run("Diagnostics with invalid ID", func(t *testing.T) {
+		_, err := client.StackDeploymentSteps.Diagnostics(ctx, "invalid-id")
+		require.Error(t, err)
+	})
+
+	t.Run("Artifacts with valid artifact name (plan-description)", func(t *testing.T) {
+		rawBytes, err := client.StackDeploymentSteps.Artifacts(ctx, step.ID, StackDeploymentStepArtifactPlanDescription)
+		assert.NoError(t, err)
+
+		b, err := io.ReadAll(rawBytes)
+		assert.NoError(t, err)
+		assert.NotEmpty(t, string(b))
+	})
+
+	t.Run("Artifacts with invalid artifact name", func(t *testing.T) {
+		_, err := client.StackDeploymentSteps.Artifacts(ctx, step.ID, "invalid-artifact-name")
+		assert.Error(t, err)
+	})
+
+	t.Run("Artifacts with invalid step ID", func(t *testing.T) {
+		_, err := client.StackDeploymentSteps.Artifacts(ctx, "invalid-id", StackDeploymentStepArtifactPlanDescription)
+		require.Error(t, err)
+	})
 }